airsm_restapi/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
dnspython==2.1.0
filelock==3.0.12
Flask==2.0.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.34.1
gunicorn==20.1.0
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.11
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
nodeenv==1.6.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
paho-mqtt==1.5.1
parso==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.10
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyJWT==2.1.0
pylint==2.6.0
python-dateutil==2.8.1
python-etcd==0.4.5
python3-openid==3.2.0
pytz==2020.5
pyzmq==21.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.2
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
